refactor/rename editisdonecommand and editdonecommand
