+++ src/main/java/seedu/task/logic/parser/EditIsDoneParser.java
@@ -5,7 +5,7 @@ import static seedu.task.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import java.util.Optional;
 
 import seedu.task.logic.commands.Command;
-import seedu.task.logic.commands.EditIsDoneCommand;
+import seedu.task.logic.commands.DoneCommand;
 import seedu.task.logic.commands.IncorrectCommand;
 
 
@@ -14,9 +14,9 @@ public class EditIsDoneParser {
     public Command parse(String args) {
         Optional<Integer> index = ParserUtil.parseIndex(args);
         if (!index.isPresent()) {
-            return new IncorrectCommand(String.format(MESSAGE_INVALID_COMMAND_FORMAT, EditIsDoneCommand.MESSAGE_USAGE));
+            return new IncorrectCommand(String.format(MESSAGE_INVALID_COMMAND_FORMAT, DoneCommand.MESSAGE_USAGE));
         }
 
-        return new EditIsDoneCommand(index.get());
+        return new DoneCommand(index.get());
     }
 }
