+++ src/main/java/seedu/task/logic/parser/EditUnDoneParser.java
@@ -5,7 +5,7 @@ import static seedu.task.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import java.util.Optional;
 
 import seedu.task.logic.commands.Command;
-import seedu.task.logic.commands.EditUnDoneCommand;
+import seedu.task.logic.commands.UnDoneCommand;
 import seedu.task.logic.commands.IncorrectCommand;
 
 
@@ -14,9 +14,9 @@ public class EditUnDoneParser {
     public Command parse(String args) {
         Optional<Integer> index = ParserUtil.parseIndex(args);
         if (!index.isPresent()) {
-            return new IncorrectCommand(String.format(MESSAGE_INVALID_COMMAND_FORMAT, EditUnDoneCommand.MESSAGE_USAGE));
+            return new IncorrectCommand(String.format(MESSAGE_INVALID_COMMAND_FORMAT, UnDoneCommand.MESSAGE_USAGE));
         }
 
-        return new EditUnDoneCommand(index.get());
+        return new UnDoneCommand(index.get());
     }
 }
